@@ -1,6 +1,6 @@
 [project]
-# Support Python 3.8+.
-requires-python = ">=3.8"
+# Support Python 3.10+.
+requires-python = ">=3.10"
 
 [tool.ruff]
 # Same as Black.
