@@ -5,8 +5,9 @@ import inspect
 import os
 import pathlib
 from collections import OrderedDict
+from collections.abc import Callable
 from enum import Enum
-from typing import Any, Callable, NewType, Tuple, Union, cast, get_args
+from typing import Any, NewType, Union, cast, get_args
 
 from .node_context import NodeContext
 from .node_data import NodeData
@@ -94,7 +95,7 @@ def eval_type(t: str | _Ty, __globals: dict[str, Any], /):
     # `compile_type_string` adds `Union`, so we need it in scope
     local_scope = {
         "Union": Union,
-        "Tuple": Tuple,
+        "Tuple": tuple,
     }
 
     try:
@@ -108,7 +109,7 @@ def union_types(types: list[_Ty]) -> _Ty:
     assert len(types) > 0
     t: Any = types[0]
     for t2 in types[1:]:
-        t = Union[t, cast(Any, t2)]
+        t = t | cast(Any, t2)
     return t
 
 
@@ -182,7 +183,7 @@ def validate_return_type(return_type: _Ty, node: NodeData):
                 f"Return type '{return_type}' must have the same number of arguments as there are outputs"
             )
 
-        for o, return_arg in zip(outputs, return_args):
+        for o, return_arg in zip(outputs, return_args, strict=False):
             if o.associated_type is not None and not is_subset_of(
                 return_arg, o.associated_type
             ):
@@ -268,7 +269,7 @@ def check_schema_types(
         raise CheckFailedError(
             f"Number of inputs and arguments don't match: {len(ann)=} != {len(inputs)=}"
         )
-    for (a_name, a_type), i in zip(ann.items(), inputs):
+    for (a_name, a_type), i in zip(ann.items(), inputs, strict=False):
         associated_type = i.associated_type
         if (
             associated_type is not None
