@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 import gc
-from typing import Callable
+from collections.abc import Callable
 
 import numpy as np
 import onnxruntime as ort
