@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import hashlib
 import struct
-from typing import Union
 
 from nodes.groups import optional_list_group, seed_group
 from nodes.properties.inputs import BaseInput, SeedInput
@@ -12,7 +11,7 @@ from nodes.utils.utils import ALPHABET
 
 from .. import random_group
 
-Source = Union[int, float, str, Seed]
+Source = int | float | str | Seed
 
 
 def SourceInput(label: str):
