@@ -1,7 +1,8 @@
 from __future__ import annotations
 
+from collections.abc import Callable, Iterable
 from dataclasses import dataclass
-from typing import Callable, Generic, Iterable, TypeVar
+from typing import Generic, TypeVar
 
 I = TypeVar("I")
 L = TypeVar("L")
