@@ -1,8 +1,9 @@
 from __future__ import annotations
 
+from collections.abc import Callable, Mapping
 from dataclasses import dataclass
 from enum import Enum
-from typing import Any, Callable, Generic, Mapping, Protocol, TypeVar
+from typing import Any, Generic, Protocol, TypeVar
 
 import navi
 
